Env + If rol Statements

+++ VistaExamenPlanner/Controllers/ExamenLokaal.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using MySql.Data.MySqlClient;
+using System.Security.Claims;
 using VistaExamenPlanner.Handler;
 using VistaExamenPlanner.Objecten;
 
@@ -20,6 +21,13 @@ namespace VistaExamenPlanner.Controllers
         [HttpGet("GetLokaal")]
         public string GetLokaal()
         {
+            string Rol = User.FindFirstValue("Rol");
+            if (Rol == "" || Rol != "3")
+            {
+                _logger.Log(LogLevel.Warning, $"GetLokaal: Someone without the rights tried to get a list of all the classrooms.");
+                return "NoPerm";
+            }
+
             using (DatabaseHandler database = new())
             {
                 MySqlCommand command = new();
@@ -39,6 +47,13 @@ namespace VistaExamenPlanner.Controllers
                 throw new ArgumentNullException(nameof(lokaal));
             }
 
+            string Rol = User.FindFirstValue("Rol");
+            if (Rol == "" || Rol != "3")
+            {
+                _logger.Log(LogLevel.Warning, $"AddLokaal: Someone without the rights tried to create a ClassRoom.");
+                return;
+            }
+
             using (DatabaseHandler database = new())
             {
                 MySqlCommand command = new();
@@ -50,11 +65,26 @@ namespace VistaExamenPlanner.Controllers
         }
 
         [Authorize]
-        [HttpPost("DeleteLokaal")]
+        [HttpDelete("DeleteLokaal")]
         public void DeleteLokaal(int IdToDropTable)
         {
+
+            string Rol = User.FindFirstValue("Rol");
+            if (Rol == "" || Rol != "3")
+            {
+                _logger.Log(LogLevel.Warning, $"DeleteLokaal: Someone without the rights tried to Delete a ClassRoom.");
+                return;
+            }
+
             using (DatabaseHandler database = new())
             {
+
+                MySqlCommand DropFromAgenda = new MySqlCommand();
+
+                DropFromAgenda.CommandText = $"DELETE FROM AgendaItem WHERE Lokaal_Id = @IdToDropTable;";
+                DropFromAgenda.Parameters.AddWithValue("@IdToDropTable", IdToDropTable);
+                database.Delete(DropFromAgenda);
+
                 MySqlCommand command = new MySqlCommand();
 
                 command.CommandText = $"DELETE FROM Lokaal WHERE Id = @IdToDropTable;";
